requirements.txt
streamlit
pypdf
python-docx
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
google-generativeai
